pyproject.toml
[tool.poetry]
name = "opennet-multithreading"
version = "0.1.0"
description = ""
authors = ["Terry <boy82530ten@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
